[#24] Refactor and set up base application class
